Add script to rotates circle when a key label is clicked
Rotate circle upon click on a key label so that the label goes to the
top-most, center position on the circle. Rotate in the diretion of the
shortest route (i.e. if the label is on the right side, the circle
rotates counterclockwise, if the label is one the left side, or at the
bottom, the circle rotates clockwise).